Fix: Moved the SoftsEncoder to _encoder directory
