+++ pytorch_forecasting/layers/_blocks/__init__.py
@@ -1,11 +1,9 @@
 from pytorch_forecasting.layers._blocks._residual_block_dsipts import ResidualBlock
 from pytorch_forecasting.layers._blocks._softs_block import (
-    SOFTSEncoderLayer,
     STADModule,
 )
 
 __all__ = [
     "ResidualBlock",
     "STADModule",
-    "SOFTSEncoderLayer",
 ]
